correct uint64_t bug

+++ include/json/config.h
@@ -5,6 +5,7 @@
 
 #ifndef JSON_CONFIG_H_INCLUDED
 # define JSON_CONFIG_H_INCLUDED
+#include <stdint.h>
 
 /// If defined, indicates that json library is embedded in CppTL library.
 //# define JSON_IN_CPPTL 1
@@ -85,8 +86,8 @@ namespace Json {
    typedef __int64 Int64;
    typedef unsigned __int64 UInt64;
 #  else // if defined(_MSC_VER) // Other platforms, use long long
-   typedef long long int Int64;
-   typedef unsigned long long int UInt64;
+   typedef int64_t Int64;
+   typedef uint64_t UInt64;
 #  endif // if defined(_MSC_VER)
    typedef Int64 LargestInt;
    typedef UInt64 LargestUInt;
